requirements.txt
# requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
aiofiles>=23.2.1
pydantic>=2.4.2
python-dotenv>=1.0.0
aiohttp>=3.8.6
yeelight>=0.7.13